ライブラリファイルの作成を改良

+++ sources/Mesh.h
@@ -33,7 +33,7 @@
 	PUBLIC void Apply() {
 		D3D11_BUFFER_DESC vertexBufferDesc = {};
 		vertexBufferDesc.Usage = D3D11_USAGE_DEFAULT;
-		vertexBufferDesc.ByteWidth = sizeof(Vertex) * vertexes.size();
+		vertexBufferDesc.ByteWidth = static_cast<UINT>(sizeof(Vertex) * vertexes.size());
 		vertexBufferDesc.BindFlags = D3D11_BIND_VERTEX_BUFFER;
 		vertexBufferDesc.CPUAccessFlags = 0;
 		D3D11_SUBRESOURCE_DATA vertexSubresourceData = {};
@@ -47,7 +47,7 @@
 
 		D3D11_BUFFER_DESC indexBufferDesc = {};
 		indexBufferDesc.Usage = D3D11_USAGE_DEFAULT;
-		indexBufferDesc.ByteWidth = sizeof(int) * indexes.size();
+		indexBufferDesc.ByteWidth = static_cast<UINT>(sizeof(int) * indexes.size());
 		indexBufferDesc.BindFlags = D3D11_BIND_INDEX_BUFFER;
 		indexBufferDesc.CPUAccessFlags = 0;
 		D3D11_SUBRESOURCE_DATA indexSubresourceData = {};
@@ -63,7 +63,7 @@
 		App::GetGraphicsContext().VSSetConstantBuffers(0, 1, &constantBuffer);
 		App::GetGraphicsContext().PSSetConstantBuffers(0, 1, &constantBuffer);
 
-		App::GetGraphicsContext().DrawIndexed(indexes.size(), 0, 0);
+		App::GetGraphicsContext().DrawIndexed(static_cast<UINT>(indexes.size()), 0, 0);
 	}
 	PUBLIC static Mesh CreateQuad() {
 		Mesh mesh;
